Add sort_date attribute in database.
Recent images will be returned by ordering based on the sort_date attribute, rather than the capture_time. This is because the capture time is revealed only in the header file, and we might want to retrieve a new image before that is uploaded. 

If a database element is being created by an image file, sort_date is set to the upload time of that image. If a database element is being modified by an image upload, the sort_date does not change. If an element is being created or updated by a header file, then the capture_date value from the header is used to set  both the capture_date attribute and the sort_date attribute in the database.